requirements.txt
djangorestframework
django-admin
Pillow  